chore: Move metadata to new file and add comments
